[DOCSET] use Matches.svelte for network vis
